// api/interface/PrebookInterface.tsx
export interface PrebookInterface {
  boardChanged: boolean;
  cancellationChanged: boolean;
  commission: number;
  currency: string;
  hotelId: string;
  isPackageRate: null;
  offerId: string;
  prebookId: string;
  price: number;
  priceDifferencePercent: number;
  priceType: string;
  secretKey: string;
  suggestedSellingPrice: number;
  supplier: string;
  supplierId: number;
  termsAndConditions: string;
  transactionId: string;
}